% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTT.R
\name{BuildTT}
\alias{BuildTT}
\title{Builds a TideTable model}
\usage{
BuildTT(data, otz = 1, start_date, start_time, end_date, end_time)
}
\arguments{
\item{data}{the data frame with observation date, observation time and height.}

\item{otz}{time zone of the observations. Default is 1 (UTC + 1)}

\item{start_date}{The start date}

\item{start_time}{The start time}

\item{end_date}{The end date}

\item{end_time}{The end time}
}
\value{
Returns a object of class "tidetable"
}
\description{
Builds a TideTable model of class "tidetable".
}
