% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{ComputeAfunc}
\alias{ComputeAfunc}
\title{#' Returns predictor vector for design matrix}
\usage{
ComputeAfunc(omega = NULL, xi = NULL)
}
\arguments{
\item{omega}{The return value of FindOmega().}

\item{xi}{Transit index}
}
\value{
A list with the selected angular velocities, their ranks and the predictor vector (Values between -1, 1).
}
\description{
Returns predictor vector for design matrix from 39 astronomical angular velocities.
}
