% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{FindOmega}
\alias{FindOmega}
\title{Returns omegas and their ranks.}
\usage{
FindOmega(tdiff)
}
\arguments{
\item{tdiff}{Length of input time series.}
}
\value{
A list with the selected angular velocities and their ranks.
}
\description{
Returns omegas and their ranks from 39 astronomical angular velocities.
}
