% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{Funcs}
\alias{Funcs}
\title{Returns predictor vector for design matrix from 44 astronomical angular velocities.}
\usage{
Funcs(xi, ma = 89, ivar = 1, tdiff)
}
\arguments{
\item{xi}{Transit index}

\item{ma}{Max number of predictors.}

\item{ivar}{Dummy.}

\item{tdiff}{Length of input time series.}
}
\value{
The predictor vector. Values between -1, 1.
}
\description{
Returns predictor vector for design matrix from 44 astronomical angular velocities.
}

