% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{NumCulm}
\alias{NumCulm}
\title{Calculates numm and k4}
\usage{
NumCulm(t, tmhwi)
}
\arguments{
\item{t}{Time in days after 1900/01/01 00:00:00 UTC.}

\item{tmhwi}{Mean high water interval (Greenwich meridian).}
}
\value{
Returns a list containing numm and k4.
}
\description{
Calculates transit number (numm) and high (1, 3) or low (2, 4) water number (k4).
}
