% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\docType{data}
\name{observation}
\alias{observation}
\title{Sample file of high and low water times and heights}
\format{
A data frame with 26819 rows and 4 variables
\describe{
  \item{observation_date}{date of tide, "yyyy/mm/dd" format, character}
  \item{observation_time}{time of tide, "hh:mm:ss" format, character}
  \item{high_or_low_water}{indication whether high (1) or low water (0) was present given date and time, integer}
  \item{height}{height of the tide, numeric}
}
}
\usage{
observation
}
\description{
A sample dataset containing observation date, time and height of high and low water
}
\keyword{datasets}
