% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{PlotTides}
\alias{PlotTides}
\title{Plot function}
\usage{
PlotTides(data)
}
\arguments{
\item{data}{Output from the TideTables function}
}
\value{
Generates eight plots
}
\description{
Plots the computed tides
}
