% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tt}
\alias{predict.tt}
\title{Synthesizes a Tide Table}
\usage{
\method{predict}{tt}(model = NULL, start_date, end_date, start_time,
  end_time, stz = 1)
}
\arguments{
\item{model}{the model you built with BuildTT()}

\item{start_date}{Start date of the Synthesis}

\item{end_date}{End date of the Synthesis}

\item{start_time}{Start time of the Synthesis}

\item{end_time}{End time of the Synthesis}
}
\value{
Returns the tide table
}
\description{
Synthesiszes a tide table
}
