% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynTT.R
\name{SynTT}
\alias{SynTT}
\title{Synthesizes a Tide Table}
\usage{
SynTT(tmodel = NULL, ssdate, sstime, sedate, setime, stz = 1)
}
\arguments{
\item{tmodel}{the model you built with BuildTT()}

\item{ssdate}{Start date of the synthesis}

\item{sstime}{End date of the synthesis}

\item{sedate}{Start time of the synthesis}

\item{setime}{End time of the synthesis}

\item{stz}{The target time zone. Defaults to UTC +1}
}
\value{
Returns a tide table as a data.table, which is identical to c.table from TideTable
}
\description{
Synthesizes a tide table, built with BuildTT()
}
