% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{Funcs}
\alias{Funcs}
\title{Returns predictor vector for design matrix}
\usage{
Funcs(tdiff, xi)
}
\arguments{
\item{tdiff}{Length of input time series.}

\item{xi}{Transit index}
}
\value{
A list with the selected angular velocities, their ranks and the predictor vector (Values between -1, 1).
}
\description{
Returns predictor vector for design matrix from 44 astronomical angular velocities.
}
